Do not merge this.
This is a really horrible first cut at a libnl-xxx higher level library to
wrap the inetdiag netlink interface.

I wrote this code just to test it out and see if the actual libnl maintainer
is interested in something of this form prior to me going down this rabbit
hole.

This is pretty useful for writing simple inetdiag apps as it is, though.